refac: transform password page to a isolated comp
