--- src/app/password/password-page.component.html
@@ -1,54 +1,0 @@
-<div class="password-page ion-padding" data-testid="password-page">
-	<div class="password-page__header ion-padding flex flex-col">
-		<span class="font-bold text-4xl">
-			{{ "PASSWORD_PAGE.TITLE" | translate }}
-		</span>
-		<span class="text-medium text-sm">
-			{{ "PASSWORD_PAGE.SUBTITLE" | translate }}
-		</span>
-	</div>
-	<div class="password-page__words flex flex-wrap w-full justify-between">
-		<div
-			class="password-page__words--word ion-padding"
-			*ngFor="let word of words; let i = index"
-		>
-			<word-box [order]="i + 1" [word]="word"></word-box>
-		</div>
-	</div>
-	<div
-		class="password-page__actions w-full flex justify-between ion-padding absolute bottom-0 left-0"
-	>
-		<div class="password_page__actions--button-group ion-padding">
-			<ion-button
-				mode="ios"
-				class="mr-2"
-				color="primary-light"
-				data-testid="password_page__actions--button-group--copy"
-				(click)="copyWordsToClipboard()"
-			>
-				<ion-icon
-					size="small"
-					slot="icon-only"
-					name="copy-outline"
-				></ion-icon>
-			</ion-button>
-			<ion-button
-				mode="ios"
-				color="primary-light"
-				data-testid="password_page__actions--button-group--save"
-			>
-				<span class="px-2">{{ "SAVE" | translate }}</span>
-			</ion-button>
-		</div>
-		<div class="password_page__actions--button-next ion-padding">
-			<ion-button
-				mode="ios"
-				color="primary-light"
-				data-testid="password_page__actions--button-group--next"
-				(click)="handleNextAction()"
-			>
-				<span class="px-5">{{ "NEXT_STEP" | translate }}</span>
-			</ion-button>
-		</div>
-	</div>
-</div>
