--- src/app/password/password-page.component.scss
@@ -1,3 +1,0 @@
-.password-page__words--word {
-	width: calc(100% / 3);
-}
