--- src/app/password/word-box/word-box.component.html
@@ -1,14 +1,0 @@
-<div
-	data-testid="word-box"
-	class="word-box border rounded-md relative py-3 pr-10"
->
-	<div
-		data-testid="word-box__order"
-		class="word-box__order inline absolute top-0 ml-3 pl-1 pr-1 text-xs text-medium"
-	>
-		{{ order }}
-	</div>
-	<span data-testid="word-box__word" class="word-box__word pl-4 font-bold">
-		{{ word }}
-	</span>
-</div>
