--- src/app/password/word-box/word-box.component.scss
@@ -1,4 +1,0 @@
-.word-box__order {
-	top: -0.5rem;
-	background: var(--ion-background-color) !important;
-}
