+++ src/app/wallets/wallet-passphrase-list/wallet-passphrase-list.component.html
@@ -1,0 +1,30 @@
+<div
+	class="wallet-passphrase-list ion-padding"
+	data-testid="wallet-passphrase-list"
+>
+	<div class="flex flex-wrap w-full justify-between">
+		<div
+			class="w-1/3 ion-padding"
+			*ngFor="let word of words; let i = index"
+		>
+			<div
+				data-testid="wallet-passphrase-list__word-box"
+				class="wallet-passphrase-list__word-box border rounded-md relative py-3 pr-10"
+				[ngClass]="{ selected: isSelected }"
+			>
+				<div
+					data-testid="wallet-passphrase-list__word-box--order"
+					class="wallet-passphrase-list__word-box--order inline absolute top-0 ml-3 pl-1 pr-1 text-xs text-medium"
+				>
+					{{ i + 1 }}
+				</div>
+				<span
+					data-testid="wallet-passphrase-list__word-box--word"
+					class="wallet-passphrase-list__word-box--word pl-4 font-bold"
+				>
+					{{ word }}
+				</span>
+			</div>
+		</div>
+	</div>
+</div>
