+++ src/app/wallets/wallet-passphrase-list/wallet-passphrase-list.component.module.ts
@@ -1,0 +1,14 @@
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+import { IonicModule } from "@ionic/angular";
+import { TranslateModule } from "@ngx-translate/core";
+
+import { PipesModule } from "@/pipes/pipes.module";
+
+import { WalletPassphraseListComponent } from "./wallet-passphrase-list.component";
+
+@NgModule({
+	declarations: [WalletPassphraseListComponent],
+	imports: [IonicModule, CommonModule, TranslateModule, PipesModule],
+})
+export class WalletPassphraseListComponentModule {}
