+++ src/app/wallets/wallet-passphrase-list/wallet-passphrase-list.component.scss
@@ -1,0 +1,6 @@
+.wallet-passphrase-list {
+	.wallet-passphrase-list__word-box--order {
+		top: -0.5rem;
+		background: var(--ion-background-color) !important;
+	}
+}
