+++ src/app/wallets/wallet-passphrase-list/wallet-passphrase-list.component.ts
@@ -1,0 +1,13 @@
+import { Component, Input } from "@angular/core";
+
+@Component({
+	selector: "wallet-passphrase-list",
+	templateUrl: "wallet-passphrase-list.component.html",
+	styleUrls: ["wallet-passphrase-list.component.scss"],
+})
+export class WalletPassphraseListComponent {
+	@Input()
+	public words: Array<string> = [];
+
+	constructor() {}
+}
