+++ src/assets/i18n/en.json
@@ -27,9 +27,12 @@
 		"DELETE_CONTACT": "Are you sure you want to delete contact '{{ name }}'?",
 		"OVERWRITE_CONTACT": "There is already a contact named '{{ name }}' with this address. Do you want to overwrite it's name with '{{ newName }}'?"
 	},
-	"PASSWORD_PAGE": {
+	"PASSPHRASE_LIST": {
 		"TITLE": "Your password",
-		"SUBTITLE": "Save this password and don't show it to anyone. This is the only way to access your wallet."
+		"SUBTITLE": "Save this password and don't show it to anyone. This is the only way to access your wallet.",
+		"CHECK_PASSPHRASE_TITLE": "Check password",
+		"CHECKPASSPHRASE_SUBTITLE": "Confirm your password. Place the password under the number to confirm.",
+		"SELECT_WORD": "Select word"
 	},
 	"COPIED_CLIPBOARD": "Copied to clipboard",
 	"COPY": "Copy",
